refactor: TeamCreatePage => Template 분리
