+++ client/src/Component/Template/TeamSetting/TeamSettingTemplate.tsx
@@ -7,7 +7,7 @@ import { TeamSettingButtonContainer } from "@Molecules/.";
 import { TeamInfo } from "@Organism/.";
 import { teamState } from "@Recoil/TeamData";
 import { userState } from "@Recoil/UserData";
-import { ProfileList } from "./Profile/ProfileList";
+import { ProfileList } from "../Profile/ProfileList";
 
 const TeamSettingTemPlateStyle = css`
   display: flex;
@@ -64,7 +64,7 @@ export const TeamSettingTemplate: React.FC = () => {
 
   return (
     <div css={TeamSettingTemPlateStyle}>
-      <TeamInfo setLocSelected={setLocSelected} teamNameRef={teamNameRef} teamInfoRef={teamInfoRef} />
+      {/* <TeamInfo setLocSelected={setLocSelected} teamNameRef={teamNameRef} teamInfoRef={teamInfoRef} /> */}
       <ProfileList handleProfileListContainer={() => {}} datas={teamInfoState?.member} person={1} profileRef={profileRef} />
       <TeamSettingButtonContainer clickUpdateButton={clickUpdateButton} clickExitButton={clickExitButton} />
     </div>
