+++ client/src/Component/Template/index.tsx
@@ -1,0 +1,2 @@
+export * from "./TeamSetting/TeamCreateTemplate";
+export * from "./TeamSetting/TeamSettingTemplate";
